# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API de Gestão de Projetos",
        "description": "Documentação da API para o sistema de gestão de projetos de energia solar",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/instalacoes": {
            "get": {
                "tags": [
                    "Instalações"
                ],
                "summary": "Listar todas as instalações",
                "description": "Retorna uma lista de todas as instalações.",
                "operationId": "5e4b0704f267aeff792ebb63ed011c39",
                "responses": {
                    "200": {
                        "description": "Lista de instalações",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Instalacao"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Instalações"
                ],
                "summary": "Criar uma nova instalação",
                "description": "Cria uma nova instalação com os dados fornecidos.",
                "operationId": "bb912ee668dd89d3d0f189fe5812bdcf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao"
                                ],
                                "properties": {
                                    "descricao": {
                                        "description": "Descrição da instalação",
                                        "type": "string",
                                        "example": "Instalação Fotovoltaica"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Instalação criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Instalacao"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar instalação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar instalação"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/instalacoes/{id}": {
            "get": {
                "tags": [
                    "Instalações"
                ],
                "summary": "Obter detalhes de uma instalação",
                "description": "Retorna os detalhes de uma instalação específica baseado no ID fornecido.",
                "operationId": "221adb02fe903c4de5948416e20b8b47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da instalação",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da instalação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Instalacao"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Instalação não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Instalação não encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Instalações"
                ],
                "summary": "Atualizar uma instalação",
                "description": "Atualiza os dados de uma instalação existente baseado no ID fornecido.",
                "operationId": "632d65ef7ee97e6a1acfa21c06028eb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da instalação a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao"
                                ],
                                "properties": {
                                    "descricao": {
                                        "description": "Descrição da instalação",
                                        "type": "string",
                                        "example": "Instalação Atualizada"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Instalação atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Instalacao"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Instalação não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Instalação não encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar instalação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao atualizar instalação"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Instalações"
                ],
                "summary": "Deletar uma instalação",
                "description": "Deleta uma instalação existente baseado no ID fornecido.",
                "operationId": "2d7052dbb89edbe9fd33593d8eb5c3da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da instalação a ser deletada",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instalação deletada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Instalação deletada com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Instalação não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Instalação não encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao deletar instalação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao deletar instalação"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Instalacao": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "descricao": {
                        "type": "string",
                        "example": "Instalação Fotovoltaica"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-29T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-29T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}